+++ benchmark/consoleTransport.js
@@ -2,18 +2,49 @@
 
 /* eslint no-console: "off", no-empty-function: "off" */
 
-const zerg = require('../src');
+const zerg = require('src');
 
 const demos = [
-    'drone', 'overlord', 'overseer', 'changeling', 'zergling', 'baneling', 'roach', 'ravager', 'hydralisk',
-    'lurker', 'swarm host', 'locust', 'queen', 'mutalisk', 'guardian', 'devourer', 'corruptor', 'brood lord', 'viper',
-    'scourge', 'defiler', 'queen', 'infestor', 'ultralisk', 'omegalisk', 'pigalisk', 'brutalisk', 'leviathan',
-    'broodling', 'infested terran', 'infested colonist', 'infested marine', 'aberration'
+  'drone',
+  'overlord',
+  'overseer',
+  'changeling',
+  'zergling',
+  'baneling',
+  'roach',
+  'ravager',
+  'hydralisk',
+  'lurker',
+  'swarm host',
+  'locust',
+  'queen',
+  'mutalisk',
+  'guardian',
+  'devourer',
+  'corruptor',
+  'brood lord',
+  'viper',
+  'scourge',
+  'defiler',
+  'queen',
+  'infestor',
+  'ultralisk',
+  'omegalisk',
+  'pigalisk',
+  'brutalisk',
+  'leviathan',
+  'broodling',
+  'infested terran',
+  'infested colonist',
+  'infested marine',
+  'aberration',
 ];
 
-zerg.enable(demos.map(function (val) {
-    return '-' + val + '*'
-}));
+zerg.enable(
+  demos.map(function (val) {
+    return '-' + val + '*';
+  })
+);
 
 var logs = [];
 logs.push(zerg.module('api'));
@@ -25,7 +56,6 @@ logs.push(zerg.module('benchmark:v1'));
 logs.push(zerg.module('benchmark:v2'));
 logs.push(zerg.module('benchmark:v3'));
 
-
 const CYCLES = 1000000;
 const results = [];
 
@@ -34,40 +64,40 @@ const originConsoleLog = console.log;
 console.log = function () {};
 
 for (let k = 0; k < CYCLES; k++) {
-    let log = logs[Math.floor(Math.random() * logs.length)];
-    let start = process.hrtime();
+  let log = logs[Math.floor(Math.random() * logs.length)];
+  let start = process.hrtime();
 
-    log.info(`${log.name} message`, 1, true);
+  log.info(`${log.name} message`, 1, true);
 
-    let diff = process.hrtime(start);
+  let diff = process.hrtime(start);
 
-    results.push(diff[0] * 1e9 + diff[1]);
+  results.push(diff[0] * 1e9 + diff[1]);
 }
 
 // enable console.log
 console.log = originConsoleLog;
 
 var stat = {
-    sum: null,
-    avg: null,
-    min: null,
-    max: null
+  sum: null,
+  avg: null,
+  min: null,
+  max: null,
 };
 
 // sum
 results.forEach(function (val) {
-    stat.sum += val;
+  stat.sum += val;
 });
 
 // min / max
 results.forEach((val) => {
-    if (stat.min === null || val < stat.min) {
-        stat.min = val;
-    }
+  if (stat.min === null || val < stat.min) {
+    stat.min = val;
+  }
 
-    if (stat.max === null || val > stat.max) {
-        stat.max = val;
-    }
+  if (stat.max === null || val > stat.max) {
+    stat.max = val;
+  }
 });
 
 stat.avg = Math.round(stat.sum / results.length);