+++ src/index.ts
@@ -1,0 +1,16 @@
+import Logger from './Logger';
+import LogListener from './LogListener';
+import {TLogListenerParams} from './types';
+
+const createLogger = () => {
+  return new Logger();
+};
+
+const createListener = (params: TLogListenerParams) => {
+  return new LogListener(params);
+};
+
+export default {
+  createLogger,
+  createListener,
+};