+++ src/transports/index.ts
@@ -1,0 +1,3 @@
+export {consoleBrowserColorful} from './consoleBrowserColorful';
+export {consoleBrowser} from './consoleBrowser';
+export {consoleNodeColorful} from './consoleNodeColorful';