+++ src/utils.ts
@@ -1,0 +1,3 @@
+export function objectKeys<O>(o: O) {
+  return Object.keys(o) as (keyof O)[];
+}